requirements.txt
anthropic>=0.40.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
PyGitHub>=2.4.0
pylint>=3.3.0
mypy>=1.13.0
tree-sitter>=0.23.0
tree-sitter-python>=0.23.0
pydantic>=2.9.0
rich>=13.9.0
python-dotenv>=1.0.0
gitpython>=3.1.43
semgrep>=1.90.0
